Handle use-clauses on child packages
When computing missing imports/qualifiers for use clauses
on child packages (e.g: 'use C' for an 'A.B.C' package), we
should make sure that we don't use the missing unit's fully
qualified name but only the missing part instead (e.g: 'A.B'
in that case).

Add a test for this.

For eng/ide/lal-refactor#36