Fix wrong test
The test's expected output was simply wrong: we should not
add the fully qualified name of the missing unit as the qualifier,
only the missing part.

Test has been adapted to not rely on Ada.Text_IO anymore, which has
a Text_IO rename alias, making things more confusing.
Rely on a local hierarchy of packages instead.

For eng/ide/lal-refactor#36